Fix geology lab's KerbNet access
Unlike the stock narrow-band scanner, the geology lab doesn't have an
animation that represents its scanning activity, so it shouldn't require one
for KerbNet access to work.  This change fixes the lab's KerbNet window saying
"please verify that this scanner is properly activated", and makes it possible
to actually use KerbNet from the lab.