Translated Rover Parts
Translated parts of the rover into Russian. I don't have much experience with this yet, so could you check the code for errors? Perhaps I could translate something wrong, but according to my observations, it seems like everything works and is displayed as it should. Encoding in UTF-8. 
I know that this translation cannot be implemented directly, but I hope it will help a little when localizing the mod into Russian in the future.