Russian Localization
Done. I spent a couple of hours creating and editing all the files. Also replaced the actions of the type "crew report" and so on. I did not translate the items at the very bottom by the type "usi life support" because either write lines for them each time in the localization or make localization for the mod itself, and then move the tags from it here.